Changed namespaces in moved files in Core to match the assembly name
i.e. SonarLint.VisualStudio.Core
