+++ src/Core.UnitTests/LanguageConverterTests.cs
@@ -21,8 +21,9 @@
 using System;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SonarLint.VisualStudio.Integration.UnitTests;
 
-namespace SonarLint.VisualStudio.Integration.UnitTests
+namespace SonarLint.VisualStudio.Core.UnitTests
 {
     [TestClass]
     public class LanguageConverterTests
