+++ src/Integration/Binding/ISolutionBindingOperation.cs
@@ -22,6 +22,7 @@ using System.Collections.Generic;
 using System.Threading;
 using EnvDTE;
 using SonarQube.Client.Models;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.Binding
 {
