+++ src/Integration/Binding/RuleSetInformation.cs
@@ -20,6 +20,7 @@
 
 using System;
 using Microsoft.VisualStudio.CodeAnalysis.RuleSets;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.Binding
 {
