+++ src/Integration/Connection/ConnectionWorkflow.cs
@@ -39,6 +39,7 @@ using SonarLint.VisualStudio.Integration.TeamExplorer;
 using SonarLint.VisualStudio.Progress.Controller;
 using SonarQube.Client.Helpers;
 using SonarQube.Client.Models;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.Connection
 {
