+++ src/Integration/Helpers/LanguageHelper.cs
@@ -19,6 +19,7 @@
  */
 
 using SonarQube.Client.Models;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.Helpers
 {
