+++ src/Integration/LocalServices/SolutionRuleSetsInformationProvider.cs
@@ -26,6 +26,7 @@ using System.Linq;
 using EnvDTE;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
 using SonarLint.VisualStudio.Integration.Resources;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration
 {
