+++ src/Integration/Service/DataModel/MinimumSupportedServerPlugin.cs
@@ -19,6 +19,7 @@
  */
 
 using System.Collections.Generic;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.Service.DataModel
 {
