+++ src/TestInfrastructure/Framework/ConfigurableHost.cs
@@ -26,6 +26,7 @@ using FluentAssertions;
 using SonarLint.VisualStudio.Integration.State;
 using SonarLint.VisualStudio.Integration.TeamExplorer;
 using SonarQube.Client;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
