+++ src/TestInfrastructure/Framework/ConfigurableSolutionRuleSetsInformationProvider.cs
@@ -23,6 +23,7 @@ using System.IO;
 using EnvDTE;
 using FluentAssertions;
 using SonarLint.VisualStudio.Integration.NewConnectedMode;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
