+++ src/TestInfrastructure/Framework/ConfigurableSolutionRuleStore.cs
@@ -22,6 +22,7 @@ using System.Collections.Generic;
 using FluentAssertions;
 using Microsoft.VisualStudio.CodeAnalysis.RuleSets;
 using SonarLint.VisualStudio.Integration.Binding;
+using Language = SonarLint.VisualStudio.Core.Language;
 
 namespace SonarLint.VisualStudio.Integration.UnitTests
 {
